Fixes #18962 - moved isc file parser into dhcp_common
