--- modules/dhcp_isc/isc_common.rb
@@ -1,7 +1,0 @@
-module ::Proxy::DHCP::ISC
-  module Common
-    def ip2hex ip
-      ip.split(".").map{|i| "%02x" % i }.join(":")
-    end
-  end
-end
