+++ test/dhcp/dhcp_isc_provider_interface_test.rb
@@ -1,0 +1,8 @@
+require 'test_helper'
+require 'dhcp_common/isc/omapi_provider'
+
+class IscDhcpProviderInterfaceTest < Test::Unit::TestCase
+  def test_provider_interface
+    assert_dhcp_provider_interface(Proxy::DHCP::CommonISC::IscOmapiProvider.new({}, nil))
+  end
+end
