+++ test/dhcp_isc/production_di_wirings_test.rb
@@ -1,8 +1,8 @@
 require 'test_helper'
 require 'tempfile'
 require 'dhcp_common/subnet_service'
-require 'dhcp_isc/dhcp_isc_main'
-require 'dhcp_isc/configuration_file'
+require 'dhcp_common/isc/omapi_provider'
+require 'dhcp_common/isc/configuration_file'
 require 'dhcp_isc/leases_file'
 require 'dhcp_isc/isc_state_changes_observer'
 require 'dhcp_isc/configuration_loader'
